Move idea button
