+++ front/app/containers/Admin/projects/components/NewIdeaButton/messages.ts
@@ -1,0 +1,32 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  newIdea: {
+    id: 'app.containers.AdminPage.ProjectEdit.newIdea',
+    defaultMessage: 'New idea',
+  },
+  addNewInput: {
+    id: 'app.containers.AdminPage.ProjectEdit.addNewInput',
+    defaultMessage: 'Add an input',
+  },
+  newIssue: {
+    id: 'app.containers.AdminPage.ProjectEdit.newIssue',
+    defaultMessage: 'Add an issue',
+  },
+  newQuestion: {
+    id: 'app.containers.AdminPage.ProjectEdit.newQuestion',
+    defaultMessage: 'Add a question',
+  },
+  newOption: {
+    id: 'app.containers.AdminPage.ProjectEdit.newOption',
+    defaultMessage: 'Add an option',
+  },
+  newContribution: {
+    id: 'app.containers.AdminPage.ProjectEdit.newContribution',
+    defaultMessage: 'Add a contribution',
+  },
+  newProject: {
+    id: 'app.containers.AdminPage.ProjectEdit.newProject',
+    defaultMessage: 'New Project',
+  },
+});
