+++ front/app/containers/Admin/projects/project/ideas/index.tsx
@@ -16,11 +16,11 @@ import Button from 'components/UI/Button';
 
 import { FormattedMessage } from 'utils/cl-intl';
 
+import NewIdeaButton from '../../components/NewIdeaButton';
 import messages from '../messages';
 
 import AnalysisBanner from './AnalysisBanner';
 import ownMessages from './messages';
-import NewIdeaButton from './NewIdeaButton';
 
 const defaultTimelineProjectVisibleFilterMenu = 'phases';
 const timelineProjectVisibleFilterMenus: TFilterMenu[] = [
