+++ front/app/containers/Admin/projects/project/messages.ts
@@ -50,34 +50,6 @@ export default defineMessages({
     id: 'app.containers.AdminPage.ProjectEdit.eventsTab',
     defaultMessage: 'Events',
   },
-  newProject: {
-    id: 'app.containers.AdminPage.ProjectEdit.newProject',
-    defaultMessage: 'New Project',
-  },
-  newIdea: {
-    id: 'app.containers.AdminPage.ProjectEdit.newIdea',
-    defaultMessage: 'New idea',
-  },
-  addNewInput: {
-    id: 'app.containers.AdminPage.ProjectEdit.addNewInput',
-    defaultMessage: 'Add an input',
-  },
-  newIssue: {
-    id: 'app.containers.AdminPage.ProjectEdit.newIssue',
-    defaultMessage: 'Add an issue',
-  },
-  newQuestion: {
-    id: 'app.containers.AdminPage.ProjectEdit.newQuestion',
-    defaultMessage: 'Add a question',
-  },
-  newOption: {
-    id: 'app.containers.AdminPage.ProjectEdit.newOption',
-    defaultMessage: 'Add an option',
-  },
-  newContribution: {
-    id: 'app.containers.AdminPage.ProjectEdit.newContribution',
-    defaultMessage: 'Add a contribution',
-  },
   viewPublicProject: {
     id: 'app.containers.AdminPage.ProjectEdit.viewPublicProject',
     defaultMessage: 'View project',
